[nav] 페이지에 주소로 접근할 경우 ( ex. http://{URL}/weight) navigation item active 되지 않는 오류 보수
